Make quit work when Lua code is hung
femto8 SDL and Lua code runs in the same thread. This means if the
Lua code runs for a long time the SDL runtime does not pump events
which leads to a 'hung' window.

Add a Lua hook that runs every so many instructions (5,0000), pumps
the event queue, and checks for the SDL_QUIT event.

Remove getchar in lua_print_error as it stops the event loop from
running.